@@ -19,9 +19,9 @@
 		"sort-packages": true
 	},
 	"require": {
-		"php": ">= 7.3",
+		"php": ">= 7.4",
 		"aipng/value-objects": "^2.1",
-		"consistence/consistence": "^2.0",
+		"myclabs/php-enum": "^1.8",
 		"nette/utils": "^2.5 || ^3.0"
 	},
 	"require-dev": {
